Struct ckb_std::dynamic_loading::CKBDLContext

#[repr(C)]#[repr(align(4096))]pub struct CKBDLContext<T>(_);

Dynamic loading context T represent a buffer type, for example: [u8; 64 * 1024], the size of T must aligned with PAGE_SIZE 4096.

Methods

impl<T> CKBDLContext<T>

pub fn new() -> Self

Create instance of dynamic loading context

pub fn load<'a>(
&'a mut self,
dep_cell_data_hash: &[u8]
) -> Result<Library, Error>

Load a shared library from dep cells See module level documentation for details

pub fn load_with_offset<'a>(
&'a mut self,
dep_cell_data_hash: &[u8],
offset: usize,
size: usize
) -> Result<Library, Error>

Load a shared library from dep cells to specified buffer offset. See module level documentation for details

This function is used for loading multiple libraries.

Example

let mut context = CKBDLContext::<[u8; 64 * 1024]>::new();
let mut size = size_of_val(&context);
let mut offset = 0;

let lib1 = context.load(&CODE_HASH_LIB_1).expect("load shared lib");
size -= lib1.consumed_size()
offset += lib1.consumed_size()

let lib2 = context.load_with_offset(&CODE_HASH_LIB_2, offset, size).expect("load shared lib");
size -= lib2.consumed_size()
offset += lib2.consumed_size()

let lib3 = context.load_with_offset(&CODE_HASH_LIB_3, offset, size).expect("load shared lib");