Function ckb_std::syscalls::load_cell_code

pub fn load_cell_code(
    buf_ptr: *mut u8, 
    len: usize, 
    content_offset: usize, 
    content_size: usize, 
    index: usize, 
    source: Source
) -> Result<usize, SysError>

Load cell code, read cell data as executable code

Return the loaded data length or a syscall error

Arguments

buf_ptr - a writable pointer used to receive the data

len - length that the buf_ptr can receives.

content_offset - offset

content_size - read length

index - index

source - source